requirements.txt
aiohttp
pytest-aiohttp
pytest-cov
pytest
pyyaml
python-dateutil
aiopg
SQLAlchemy
uvloop
aioredis
psycopg2-binary
pytz
aiojobs
argon2_cffi
ciso8601
marshmallow>==3.0
async-property
ujson